+++ tests/test_utils.py
@@ -2,7 +2,11 @@ from pathlib import Path
 
 import pytest
 
-from syrupy.utils import walk_snapshot_dir
+from syrupy.exceptions import FailedToLoadModuleMember
+from syrupy.utils import (
+    import_module_member,
+    walk_snapshot_dir,
+)
 
 
 def makefiles(testdir, filetree, root=""):
@@ -39,3 +43,25 @@ def test_walk_dir_skips_non_snapshot_path(testfiles):
     assert {
         str(Path(p).relative_to(Path.cwd())) for p in walk_snapshot_dir(testdir.tmpdir)
     } == {"__snapshots__/snapfile1.ambr", "__snapshots__/snapfolder/snapfile2.svg"}
+
+
+def dummy_member():
+    return 123
+
+
+def test_import_module_member_imports_member():
+    imported_member = import_module_member(f"{__name__}.dummy_member")
+    assert imported_member() == 123
+
+
+@pytest.mark.parametrize(
+    "path",
+    [
+        "dummy_member",
+        f"{__name__}badpath.dummy_member",
+        f"{__name__}.dummy_memberbadmember",
+    ],
+)
+def test_import_module_member_with_bad_path_raises_exception(path):
+    with pytest.raises(FailedToLoadModuleMember):
+        import_module_member(path)