+++ CHANGELOG.md
@@ -9,6 +9,7 @@ From v1.0.0 onwards, this project adheres to [Semantic Versioning](https://semve
 ## Master (Unreleased)
 
 - Conversion of all `os.path` and `os.walk` calls to use `pathlib` instead, setting `pathlib` as the new preferred way of doing path operations (#130)
+- Add `--snapshot-default-extension` option to specify extension class via pytest cli (#132)
 - Fix bug where snapshot diffs were erroneously printed (#135)
 - Fix bug where snapshot names were incorrectly matching tests (#136)
 - Fix bug where deleted snapshots where incorrectly colored (#136)